Added ogles_gpgpu::VideoSource class
This provides similar functionality to ogles_gpgpu::Core, but it is intended to work with a single linked pipeline of filters, rather than the top down processing that is used by ogles_gpgpu::Core.  There is also no default display.  If the user wants a display, they must add that explicitly to the fitler chain.  It serves as an input image source to the filter chain and provides a convenience operator() that allows processing of image buffers (including iOS CvPixelBufferRef pointers), and pre-existing textures.

Added initial FIFO filter (for time N to N+1 comparisons).

Added an initial two input filter.